feat: Add ClientOnly component for client-side rendering
feat: Implement useAuthState hook for managing authentication state

feat: Create useBackgroundUpdates hook for background data updates

feat: Develop useCacheDebug hook for debugging caching issues

feat: Introduce useClientEffect hook for executing code on client side

feat: Implement useInitialCache hook for initial cache loading

feat: Create useIsBrowser hook for detecting browser environment

feat: Develop useLocalStorage hook for safe localStorage access

feat: Implement useProjectFetching hook for handling project fetching logic

feat: Create useProjectOperations hook for managing project operations

feat: Develop useProjectRefresh hook for refreshing project data

feat: Implement useThemeManager hook for theme management with system preference support

feat: Create useTierTracker hook for tracking user tier changes
