+++ src/components/ClientOnly/index.tsx
@@ -1,0 +1,1 @@
+export { ClientOnly } from "./ClientOnly";
