+++ src/hooks/use-local-storage.ts
@@ -1,0 +1,55 @@
+import { useState, useEffect } from "react";
+import { useIsBrowser } from "./use-is-browser";
+
+/**
+ * Hook for safely accessing localStorage with SSR support
+ * @param key The localStorage key to access
+ * @param initialValue The initial value to use
+ * @returns [storedValue, setValue] tuple similar to useState
+ */
+export function useLocalStorage<T>(
+  key: string,
+  initialValue: T
+): [T, (value: T | ((val: T) => T)) => void] {
+  // State to store our value
+  const [storedValue, setStoredValue] = useState<T>(initialValue);
+
+  // Check if we're in the browser environment
+  const isBrowser = useIsBrowser();
+
+  // Initialize on mount or when key changes
+  useEffect(() => {
+    if (isBrowser) {
+      try {
+        const item = localStorage.getItem(key);
+        const value = item ? JSON.parse(item) : initialValue;
+        setStoredValue(value);
+      } catch (error) {
+        console.error(`Error reading localStorage key "${key}":`, error);
+        setStoredValue(initialValue);
+      }
+    }
+  }, [isBrowser, key, initialValue]);
+
+  // Return a wrapped version of useState's setter function that
+  // persists the new value to localStorage
+  const setValue = (value: T | ((val: T) => T)) => {
+    try {
+      // Allow value to be a function so we have same API as useState
+      const valueToStore =
+        value instanceof Function ? value(storedValue) : value;
+
+      // Save state
+      setStoredValue(valueToStore);
+
+      // Save to localStorage
+      if (isBrowser) {
+        localStorage.setItem(key, JSON.stringify(valueToStore));
+      }
+    } catch (error) {
+      console.error(`Error setting localStorage key "${key}":`, error);
+    }
+  };
+
+  return [storedValue, setValue];
+}
